fix for macos usize / u64 type mismatches (#140)
* fix for macos pixelbuffer dimension types

* size_t_is_usize

* update core-media

* lint fix

* simplification

+++ macos/core-media/src/format_description.rs
@@ -74,7 +74,7 @@ impl VideoFormatDescription {
         unsafe {
             let mut ret = std::ptr::null();
             let pointers: Vec<_> = parameter_sets.iter().map(|&ps| ps.as_ptr()).collect();
-            let sizes: Vec<_> = parameter_sets.iter().map(|ps| ps.len() as u64).collect();
+            let sizes: Vec<_> = parameter_sets.iter().map(|ps| ps.len()).collect();
             result(
                 sys::CMVideoFormatDescriptionCreateFromH264ParameterSets(
                     std::ptr::null_mut(),
